% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areal_apportionment.R
\name{areal_apportionment}
\alias{areal_apportionment}
\title{EJ Population Areal Apportionment}
\usage{
areal_apportionment(
  ejscreen_bgs_data,
  facility_buff,
  facil_data,
  path_raster_layer
)
}
\arguments{
\item{ejscreen_bgs_data}{EJSCREEN data}

\item{facility_buff}{Polygons representing buffered areas of interest}

\item{facil_data}{Original facility-level data (for non-buffered shapes)}

\item{path_raster_layer}{}
}
\value{

}
\description{
Function to apportion census block group population using dasymetric raster data
when using intersection approach to creating buffers around areas of interest.
This function currently uses 1kmX1km raster data from NASA's Socioeconomic
Data and Applications Center (SEDAC)
}
