% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJCountTable.R
\name{EJCountTable}
\alias{EJCountTable}
\title{EJ Count Table}
\usage{
EJCountTable(input.data, save.option = F)
}
\arguments{
\item{save.option}{}
}
\value{

}
\description{
Creates a two-way table that summaries the number and percentage of locations
to examine overlap in demographic and environmental indicators of concern.
}
\examples{
count.table <- EJCountTable(input.data = z, save.option = F)
}
