% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJdemographics.R
\name{EJdemographics}
\alias{EJdemographics}
\title{EJ Demographics}
\usage{
EJdemographics(data, gis_method, buffer, threshold)
}
\arguments{
\item{gis_method}{User specified method of creating buffers around areas of interest (intersect, centroid, intersection).}

\item{buffer}{Distance(s) used to create buffers.}
}
\value{

}
\description{
Creates boxplots that display distribution of demographic characteristics of
potential concern in areas of interest.
}
