% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJIndexes.R
\name{EJIndexes}
\alias{EJIndexes}
\title{EJ Indexes}
\usage{
EJIndexes(data, gis_method, buffer)
}
\arguments{
\item{gis_method}{User specified method of creating buffers around areas of interest (intersect, centroid, intersection).}

\item{buffer}{Distance(s) used to create buffers.}
}
\value{

}
\description{
Creates boxplots that display distribution of EJ indexes of
potential concern in areas of interest.
}
