% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJMaps.R
\name{EJMaps}
\alias{EJMaps}
\title{EJ Maps}
\usage{
EJMaps(
  input.data,
  indic.option = NULL,
  perc.geog = NULL,
  input.name = NULL,
  save.options = F
)
}
\arguments{
\item{save.options}{}
}
\value{

}
\description{
Produces an interactive map that displays all facilities. Facility names (user
defined) are displayed for each facility. Use national or state percentiles (default: US)
}
\examples{
maps <- EJMaps(input.data = z, perc.geog = 'US', save.option = F)
}
