% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen.download.gdb.R
\name{ejscreen.download.local}
\alias{ejscreen.download.local}
\title{EJSCREEN data download}
\usage{
ejscreen.download.local(
  folder = "~/data/",
  file,
  yr = NULL,
  ftpurlbase = "https://gaftp.epa.gov/EJSCREEN/",
  justreadname = NULL,
  addflag = FALSE,
  cutoff = 80,
  or.tied = TRUE
)
}
\arguments{
\item{or.tied}{}
}
\value{

}
\description{
Downloads most recent data from EJSCREEN. Code for this function was borrowed
from Mark Corrales' ejscreen function on github at ejanalysis/ejscreen.
}
