% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_download.R
\name{ejscreen_download}
\alias{ejscreen_download}
\title{Support function for EJSCREEN data download}
\usage{
ejscreen_download(
  folder = "EJSCREEN data",
  file,
  yr = NULL,
  ftpurlbase = "https://gaftp.epa.gov/EJSCREEN/",
  state = NULL,
  justreadname = NULL,
  addflag = FALSE,
  cutoff = 80
)
}
\arguments{
\item{state}{}
}
\value{
EJSCREEN in data.frame/sf format.
}
\description{
Downloads most data from EJSCREEN ftp. Code for this function was adapted
from Mark Corrales' ejscreen function on github at ejanalysis/ejscreen.
}
