% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJSCREENBufferAPI.R
\name{EJSCREENBufferAPI}
\alias{EJSCREENBufferAPI}
\title{Support function to call EJSCREEN API}
\usage{
EJSCREENBufferAPI(input_data, dist)
}
\arguments{
\item{input_data}{facility location data as sf data.frame (point, (multi-)linestring, (multi-)polygon).}

\item{dist}{a numeric value to designate the buffer distance in miles.}
}
\value{
a tidy data.frame
}
\description{
This is a canned loop function for calling the EJSCREEN API and returning the fire-hose of data as a tidy data.frame (one row per input coordinate/shape).
}
