% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJSCREENBufferBlocks.R
\name{EJSCREENBufferBlocks}
\alias{EJSCREENBufferBlocks}
\title{Method for using census blocks for buffering in EJ proximity analysis.}
\usage{
EJSCREENBufferBlocks(dta_year, buff_dta, ejscreen_bgs_data, ejvarlist)
}
\arguments{
\item{dta_year}{The data vintage year specified by user}

\item{buff_dta}{An sf data.frame containing the (buffered) LOIs}

\item{ejscreen_bgs_data}{EJSCREEN data in sf data.frame format.}

\item{ejvarlist}{Meta-data relevant to running the screening analysis}
}
\value{
A list containing 2 sublists: (1) a LOI-buffer-level screening summary, and (2) a CBG-level screening summary for all block groups within affected communities.
}
\description{
Method for using census blocks for buffering in EJ proximity analysis.
}
