% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJSCREENBufferRaster.R
\name{EJSCREENBufferRaster}
\alias{EJSCREENBufferRaster}
\title{Method for using raster for buffering in EJ proximity analysis.}
\usage{
EJSCREENBufferRaster(ejscreen_bgs_data, facility_buff, ejvarlist, year)
}
\arguments{
\item{ejscreen_bgs_data}{EJSCREEN data in sf data.frame format.}

\item{facility_buff}{An sf data.frame containing the (buffered) LOIs}

\item{ejvarlist}{Meta-data relevant to running the screening analysis}

\item{year}{The data vintage year specified by user}
}
\value{
A list containing 2 sublists: (1) a LOI-buffer-level screening summary, and (2) a CBG-level screening summary for all block groups within affected communities.
}
\description{
Method for using raster for buffering in EJ proximity analysis.
}
