% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJWaterReturnCatchmentBuffers.R
\name{EJWaterReturnCatchmentBuffers}
\alias{EJWaterReturnCatchmentBuffers}
\title{EJ Return Water Catchment Buffers}
\usage{
EJWaterReturnCatchmentBuffers(
  input_data,
  ds_us_mode,
  ds_us_dist,
  buff_dist,
  input_type,
  attains
)
}
\arguments{
\item{ds_us_mode}{Option for upstream or downstream. Default is downstream.}

\item{ds_us_dist}{Distance up/downstream. Default is 50 miles.}

\item{buff_dist}{Distance in miles to buffer out. Default is 1 mile.}

\item{input_type}{Type of data inputted. Options limited to 'sf' and 'catchment'}

\item{attains}{}
}
\value{

}
\description{
This function returns:
(1) a SF data.frame containing user-specified buffers around catchments
(2) a data.table containing raw ATTAINs API data (OPTIONAL)
}
\examples{

}
