% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_acs_data.R
\name{fetch_acs_data}
\alias{fetch_acs_data}
\title{Support function to extract data from ACS}
\usage{
fetch_acs_data(year, state_filter = NULL)
}
\arguments{
\item{year}{Year of ACS data desired by user.}

\item{state_filter}{Users may restrict screening to a particular state in the
contiguous US. If so, users can specify a state. Default is to conduct
screening for the entire contiguous US.}
}
\value{
data.frame of national ACS demographic data at block group resolution.
}
\description{
This function looks for demographic data from ACS. First checks if this data is
in working directory. If not, it creates a directory and downloads most recent
data.
}
