% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_blockcents.R
\name{fetch_blockcents}
\alias{fetch_blockcents}
\title{Support function to download and import vintage-appropriate census block centroids}
\usage{
fetch_blockcents(year = NULL)
}
\arguments{
\item{year}{Year of data vintage}
}
\value{
Data.frame containing coordinates of census block centroids.
}
\description{
Support function to download and import vintage-appropriate census block centroids
}
