% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ejvars.R
\name{fetch_ejvars}
\alias{fetch_ejvars}
\title{Support function to compile EJSCREEN variable names for data vintage}
\usage{
fetch_ejvars(year = NULL)
}
\arguments{
\item{year}{Vintage of EJSCREEN data}
}
\value{
List of variable names and other metadata for use by EJfunction()
}
\description{
Support function to compile EJSCREEN variable names for data vintage
}
