% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_rasters.R
\name{fetch_rasters}
\alias{fetch_rasters}
\title{Support function to download and extract pop count raster data for analysis}
\usage{
fetch_rasters(year = NULL)
}
\arguments{
\item{year}{User-specified year of data vintage for analysis.}
}
\value{
A raster object
}
\description{
Support function to download and extract pop count raster data for analysis
}
