% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_dynamic_buffer.R
\name{st_dynamic_buffer}
\alias{st_dynamic_buffer}
\title{Dynamic buffering projections}
\usage{
st_dynamic_buffer(input_sf, buff_dist)
}
\arguments{
\item{input_sf}{Required. Input sf data.frame}

\item{buff_dist}{Required. Buffer distance}
}
\value{
Data.frame sf object with buffered geometry.
}
\description{
Dynamic buffering projections
}
