% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJfunction.R
\name{EJfunction}
\alias{EJfunction}
\title{Primary function that performs EJ/demographic screening.}
\usage{
EJfunction(LOI_data, data_year = NULL, buffer = NULL, raster = T, state = NULL)
}
\arguments{
\item{LOI_data}{Required. Location of interest. Location data must be an SF object to undergo screening analysis.}

\item{data_year}{Option for which EJSCREEN (and corresponding ACS) data vintage to return. Default is 2023.}

\item{buffer}{Distance(s) used to create buffers (miles). Default is 1, 3, and 5 miles for points/linestrings and 0 miles for polygons.}

\item{raster}{Block or raster apportioning for buffer? Default is T. If set to false, uses EJSCREEN's block approach.}

\item{state}{User can restrict screening to particular states. Default is to screen for entire US.}
}
\value{
A list containing 2 sublists: (1) a LOI-buffer-level screening summary, and (2) a CBG-level screening summary for all block groups within affected communities.
}
\description{
Primary function that performs EJ/demographic screening.
}
