% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJHeatTables.R
\name{EJHeatTables}
\alias{EJHeatTables}
\title{EJ Heat Tables}
\usage{
EJHeatTables(
  input.data,
  type,
  geog_lvl = NULL,
  keepid = NULL,
  topN = NULL,
  topN_dta_idx = NULL,
  input.names = NULL,
  save.option = F
)
}
\arguments{
\item{save.option}{}
}
\value{

}
\description{
Produces one of three possible heat tables.
1) Summary of median CBG value across all facilities ('all')
2) Summary of median CBG value for a single facility ('single')
3) Summary of median CBG values for "top N" facilities ('topn')
}
\details{
geog_lvl denotes whether to use nat'l/state percentiles (nat'l default): options 'US'/'state'
when type = 'single, keepid must be set by user, denotes the (integer) rowid of the facility
topN is an integer set by user when using ranking table. (Default is 10, min is 0.)
topN_dta_idx denotes which EJfunction() output data to use (e.g. when there are mult. buff. distances)
-> it is an integer. Default setting is 1
input.names is user-provided list of facility names. it must be same length as input.data
}
\examples{
y1 <- EJHeatTables(input.data = y, type = 'topn', topN = 5, topN_dta_idx = 3)
y2 <- EJHeatTables(input.data = y, type = 'all', geog_lvl = 'state')
}
