% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJHeatTables.R
\name{EJHeatTables}
\alias{EJHeatTables}
\title{EJ Heat Tables}
\usage{
EJHeatTables(
  input_data,
  heat_table_type,
  heat_table_geog_lvl = NULL,
  heat_table_keepid = NULL,
  heat_table_topN = NULL,
  save_option = F
)
}
\arguments{
\item{heat_table_type}{All facilities, single facility, top N facilities. If top N facilities, must specify topN.}

\item{heat_table_geog_lvl}{state or US}

\item{save_option}{Option to save heat table to a folder in working directory. Default is FALSE.}
}
\value{

}
\description{
Produces one of three possible heat tables.
1) Summary of median CBG value across all facilities ('all')
2) Summary of median CBG value for a single facility ('single')
3) Summary of median CBG values for "top N" facilities ('topn')
}
\details{
heat_table_geog_lvl denotes whether to use nat'l/state percentiles (nat'l default): options 'US'/'state'
when type = 'single, heat_table_keepid must be set by user, denotes the (integer) rowid of the facility
topN is an integer set by user when using ranking table. (Default is 5 for fit, min is 0, max is 10.)
}
\examples{
y1 <- EJHeatTables(input_data = y, heat_table_type = 'topn', heat_table_topN = 5)
y2 <- EJHeatTables(input_data = y, heat_table_type = 'all', heat_table_geog_lvl = 'state')
}
