% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJMaps.R
\name{EJMaps}
\alias{EJMaps}
\title{EJ Maps}
\usage{
EJMaps(input_data, indic_option = NULL, perc_geog = NULL, save_option = F)
}
\arguments{
\item{indic_option}{Options are 'total', 'environmental','demographic'. 'total' is default.}

\item{perc_geog}{Options are 'US' or 'state'.}

\item{save_option}{}
}
\value{

}
\description{
Produces an interactive map that displays all facilities. Facility names (user
defined) are displayed for each facility. Use national or state percentiles (default: US)
}
\examples{
maps <- EJMaps(input_data = z, perc_geog = 'US', save.option = F)
}
