% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJMaps.R
\name{EJMaps}
\alias{EJMaps}
\title{Produces an interactive map that displays all facilities.}
\usage{
EJMaps(
  input_data,
  indic_option = NULL,
  geography = NULL,
  threshold = 80,
  facil_name = NULL
)
}
\arguments{
\item{input_data}{Required. Screening object returned from an EJFunction() call.}

\item{indic_option}{EJSCREEN indicators to display? Options are 'total','environmental','demographic'. Default is 'total'.}

\item{geography}{National or state percentiles used in map? Options are 'state' or 'US' (default).}

\item{threshold}{Percentile threshold for map coloring. Default is 80.}

\item{facil_name}{Column name in LOI input data denoting LOI name. String only.}
}
\value{
A list containing leaflet map object(s).
}
\description{
Produces an interactive map that displays all facilities.
}
\examples{
maps <- EJMaps(input_data = z, geography = 'US', save.option = F)
}
