% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJRanking.R
\name{EJRanking}
\alias{EJRanking}
\title{EJ Ranking Function}
\usage{
EJRanking(
  input_data,
  rank_type = "location",
  rank_geography_type = "US",
  rank_count = 10,
  save_option = F
)
}
\arguments{
\item{rank_type}{Required. Either 'location' or 'CBG'}

\item{rank_geography_type}{Either 'US' or 'state}

\item{rank_count}{Required.}

\item{save_option}{}
}
\value{

}
\description{
Returns information for facilities or census block groups ranked by the number
of indicators that exceed the 80th percentile.
}
\examples{
# Facility and CBG rankings
facil.ranking <- EJRanking(input_data = a2, rank_count = 10, rank_type = 'location',
                            rank_geography_type = 'US', save_option = F)

cbg.ranking <- EJRanking(input_data = a2, rank_type = 'cbg')
}
