% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJWaterReturnCatchmentBuffers.R
\name{EJWaterReturnCatchmentBuffers}
\alias{EJWaterReturnCatchmentBuffers}
\title{EJ Return Water Catchment Buffers}
\usage{
EJWaterReturnCatchmentBuffers(input_data, ds_us_mode = "DM", ds_us_dist = 10)
}
\arguments{
\item{input_data}{Required. An sf POINT data.frame with coordinates falling in the US.}

\item{ds_us_mode}{Option for upstream or downstream. Default is downstream, main ("DM"). Other options: "DD","UT","UM".}

\item{ds_us_dist}{Distance up/downstream. Default is 10 miles.}
}
\value{
This function returns: (1) a SF data.frame containing up/downstream flowlines of a user-specified distance and (2) a list of downstream catchment IDs
}
\description{
EJ Return Water Catchment Buffers
}
