% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data_ej.R
\name{fetch_data_ej}
\alias{fetch_data_ej}
\title{Fetch data from EJSCREEN}
\usage{
fetch_data_ej(state_filter)
}
\arguments{
\item{state_filter}{Users may restrict screening to a particular state in the
contiguous US. If so, users can specify a state. Default is to conduct
screening for the entire contiguous US.}
}
\value{

}
\description{
This function looks for data from EJSCREEN. First checks if from EJSCREEN is
in working directory. If not, it creates a directory and downloads most recent
data.
}
